fix: rename connect postgres func
