--- internal/domain/auth.go
@@ -1,6 +1,0 @@
-package domain
-
-type Signup struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
