+++ internal/route/auth.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 
 	"github.com/dilyara4949/flight-booking-api/internal/config"
-	"github.com/dilyara4949/flight-booking-api/internal/controller"
+	"github.com/dilyara4949/flight-booking-api/internal/handler"
 	"github.com/dilyara4949/flight-booking-api/internal/repository"
 	"github.com/dilyara4949/flight-booking-api/internal/service"
 	"github.com/gin-gonic/gin"
@@ -13,7 +13,7 @@ import (
 func NewSignupRoute(cfg config.Config, db *sql.DB, group *gin.RouterGroup) {
 	repo := repository.NewUserRepository(db)
 	authService := service.NewAuthService(repo)
-	authController := controller.NewAuthController(authService, cfg)
+	authController := handler.NewAuthHandler(authService, cfg)
 
 	group.POST("signup", authController.Signup)
 }
